Better app layout (for the last time)
